% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hal_with_screening.R, R/screening.R
\name{aorder}
\alias{aorder}
\title{Borrowed from sl3. Internal use.}
\usage{
aorder(mat, index, along = 1)

aorder(mat, index, along = 1)
}
\description{
Borrowed from sl3. Internal use.

Borrowed from sl3. Internal use.
}
