% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_copy_map}
\alias{apply_copy_map}
\title{Apply copy map}
\usage{
apply_copy_map(X, copy_map)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}

\item{copy_map}{the copy map}
}
\description{
OR duplicate training set columns together
}
\examples{
\donttest{
gendata <- function(n, g0, Q0) {
  W1 <- runif(n, -3, 3)
  W2 <- rnorm(n)
  W3 <- runif(n)
  W4 <- rnorm(n)
  A <- rbinom(n, 1, g0(W1, W2, W3, W4))
  Y <- rbinom(n, 1, Q0(A, W1, W2, W3, W4))
  data.frame(A, W1, W2, W3, W4, Y)
}
set.seed(1234)
data <- gendata(100, g0 = g0_linear, Q0 = Q0_trig1)
covars <- setdiff(names(data), "Y")
X <- as.matrix(data[, covars, drop = FALSE])
basis_list <- enumerate_basis(X)
x_basis <- make_design_matrix(X, basis_list)
copy_map <- make_copy_map(x_basis)
pred_x_basis_uniq <- apply_copy_map(pred_x_basis, copy_map)
}
}
