% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asdgCMatrix_}
\alias{asdgCMatrix_}
\title{Fast Coercion to Sparse Matrix}
\usage{
asdgCMatrix_(XX_)
}
\arguments{
\item{XX_}{An object of class \code{Matrix} that has a sparse structure
suitable for coercion to a sparse matrix format of \code{dgCMatrix}.}
}
\value{
An object of class \code{dgCMatrix}, coerced from input \code{XX_}.
}
\description{
Fast and efficient coercion of standard matrix objects to sparse matrices.
Borrowed from http://gallery.rcpp.org/articles/sparse-matrix-coercion/.
INTERNAL USE ONLY.
}
