% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_xscale}
\alias{calc_xscale}
\title{Calculating Centered and Scaled Matrices}
\usage{
calc_xscale(X, xcenter)
}
\arguments{
\item{X}{A sparse matrix, to be centered.}

\item{xcenter}{A vector of column means to be used for centering X.}
}
\description{
Calculating Centered and Scaled Matrices
}
\keyword{internal}
