% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_knot_screening.R
\name{cluster_knot_screening}
\alias{cluster_knot_screening}
\title{Function to screen down knot-points using clustering algorithms}
\usage{
cluster_knot_screening(X, algorithm, num_clusters = NULL, max_degree = NULL)
}
\arguments{
\item{X}{An input \code{matrix} containing observations and covariates following
standard conventions in problems of statistical learning.}

\item{algorithm}{The clustering algorithm to be used. Currently, only
\code{pam} and \code{kmeans} are supported. See \section{Details} for more
information.}

\item{num_clusters}{The number of desired clusters (knot-points) for each
HAL basis function.}

\item{max_degree}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (NULL) corresponds to generating
basis functions for the full dimensionality of the input matrix.}
}
\value{
A \code{list} of basis functions generated for all covariates and
interaction thereof up to a pre-specified degree.
}
\description{
Screen down HAL knot-points using clustering algorithms to
reduce the subsequent computational and memory cost of making the augmented
design matrix and fitting HAL.
}
\examples{
gendata <- function(n) {
  W1 <- runif(n, -3, 3)
  W2 <- rnorm(n)
  W3 <- runif(n)
  W4 <- rnorm(n)
  g0 <- plogis(0.5 * (-0.8 * W1 + 0.39 * W2 + 0.08 * W3 - 0.12 * W4))
  A <- rbinom(n, 1, g0)
  Q0 <- plogis(0.15 * (2 * A + 2 * A * W1 + 6 * A * W3 * W4 - 3))
  Y <- rbinom(n, 1, Q0)
  data.frame(A, W1, W2, W3, W4, Y)
}
set.seed(1234)
data <- gendata(100)
covars <- setdiff(names(data), "Y")
X <- as.matrix(data[, covars, drop = FALSE])
screened_basis_list <- cluster_knot_screening(X, "kmeans", 10)
}
