% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{cv_lasso}
\alias{cv_lasso}
\title{Cross-validated LASSO on Indicator Bases}
\usage{
cv_lasso(x_basis, y, n_lambda = 100, n_folds = 10)
}
\arguments{
\item{x_basis}{...}

\item{y}{...}

\item{n_folds}{...}
}
\description{
Fits the LASSO regression using a customized procedure, with cross-validation
based on origami
}
