% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_basis}
\alias{evaluate_basis}
\title{Generate Basis Functions}
\usage{
evaluate_basis(basis, X, x_basis, basis_col)
}
\arguments{
\item{basis}{The basis function.}

\item{X}{The design matrix, containing the original data.}

\item{x_basis}{The HAL design matrix, containing indicator functions.}

\item{basis_col}{Numeric indicating which column to populate.}
}
\description{
Populates a column (indexed by basis_col) of x_basis with basis indicators.
}
