% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_lambda_max}
\alias{find_lambda_max}
\title{Find maximum L1 regularization constant}
\usage{
find_lambda_max(X, y, xscale)
}
\arguments{
\item{X}{...}

\item{y}{...}

\item{xscale}{...}
}
\description{
Find maximum L1 regularization constant
}
