% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{fit_hal}
\alias{fit_hal}
\title{HAL: The Highly Adaptive Lasso}
\usage{
fit_hal(X, Y, degrees = NULL, fit_type = c("glmnet", "lassi"),
  n_folds = 10, use_min = TRUE, reduce_basis = NULL,
  family = c("gaussian", "binomial"), return_lasso = FALSE,
  return_x_basis = FALSE, basis_list = NULL, lambda = NULL,
  fit_glmnet = FALSE, ..., yolo = TRUE)
}
\arguments{
\item{X}{An input \code{matrix} containing observations and covariates
following standard conventions in problems of statistical learning.}

\item{Y}{A \code{numeric} vector of obervations of the outcome variable of
interest, following standard conventions in problems of statistical learning.}

\item{degrees}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{fit_type}{The specific routine to be called when fitting the LASSO
regression in a cross-validated manner. Choosing the \code{glmnet} option
will result in a call to \code{cv.glmnet} while \code{lassi} will produce
a (faster) call to a custom LASSO routine using the \code{origami} package.}

\item{n_folds}{Integer for the number of folds to be used when splitting the
data for cross-validation. This defaults to 10 as this is the convention for
v-fold cross-validation.}

\item{use_min}{Determines which lambda is selected from \code{cv.glmnet}.
\code{TRUE} corresponds to \code{"lambda.min"} and \code{FALSE} corresponds
to \code{"lambda.1se"}.}

\item{reduce_basis}{A \code{numeric} value bounded in the open interval (0,1)
indicating the minimum proportion of 1's in a basis function column needed
for the basis function to be included in the procedure to fit the Lasso. Any
basis functions with a lower proportion of 1's than the specified cutoff
will be removed. This argument defaults to \code{NULL}, in which case all
basis functions are used in the lasso-fitting stage of the HAL algorithm.}

\item{family}{A \code{character} corresponding to the error family for a
generalized linear model. Options are limited to "gaussian" for fitting a
standard general linear model and "binomial" for logistic regression.}

\item{return_lasso}{A \code{logical} indicating whether or not to return
the \code{glmnet} fit of the lasso model.}

\item{return_x_basis}{A \code{logical} indicating whether or not to return
the matrix of (possibly reduced) basis functions used in the HAL lasso fit.}

\item{basis_list}{The full set of basis functions generated from the input
data X (via a call to \code{enumerate_basis}). The dimensionality of this
structure is dim = (n * 2^(d - 1)), where n is the number of observations
and d is the number of columns in X.}

\item{lambda}{A user-specified array of values of the lambda tuning parameter
of the Lasso L1 regression. If \code{NULL}, \code{cv.glmnet} will be used to
automatically select a CV-optimal value of this regularization parameter. If
specified, the Lasso L1 regression model will be fit via \code{glmnet},
returning regularized coefficient values for each value in the input array.}

\item{fit_glmnet}{A \code{logical} specifying whether the array of values
specified should be passed to \code{cv.glmnet} in order to pick the optimal
value (based on cross-validation) (when set to \code{FALSE}) or to simply
fit along the sequence of values (or single value) using \code{glmnet}.}

\item{...}{Other arguments passed to \code{cv.glmnet}. Please consult the
documentation for \code{glmnet} for a full list of options.}

\item{yolo}{A \code{logical} indicating whether to print one of a curated
selection of quotes from the HAL9000 computer, from the critically acclaimed
epic science-fiction film "2001: A Space Odyssey" (1968).}
}
\value{
Object of class \code{hal9001}, containing a list of basis functions,
 a copy map, coefficients estimated for basis functions, and timing results
 (for assessing computational efficiency).
}
\description{
Estimation procedure for HAL, the Highly Adaptive LASSO
}
\details{
The procedure uses a custom C++ implementation to generate a design
 matrix consisting of basis functions corresponding to covariates and
 interactions of covariates and to remove duplicate columns of indicators.
 The LASSO regression is fit to this (usually) very wide matrix using either
 a custom implementation (based on the \code{origami} package) or by a call
 to \code{cv.glmnet} from the \code{glmnet} package.
}
