% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening.R
\name{fit_hal_with_screening}
\alias{fit_hal_with_screening}
\title{The selective Highly Adaptive Lasso using MARS-based variable and interaction screening}
\usage{
fit_hal_with_screening(
  X,
  Y,
  X_unpenalized = NULL,
  max_degree = 3,
  smoothness_orders = 1,
  num_knots = ceiling(c(sqrt(length(Y)), length(Y)^(1/3), length(Y)^(1/4))),
  reduce_basis = NULL,
  family = c("gaussian", "binomial", "poisson"),
  lambda = NULL,
  id = NULL,
  weights = NULL,
  offset = NULL,
  fit_control = list(cv_select = TRUE, use_min = TRUE, lambda.min.ratio = 1e-04,
    prediction_bounds = "default"),
  screener_control = list(screen_interactions = TRUE, max_degree = max_degree,
    pruning_method = ifelse(length(Y) > 500, "backward", "cv"), family = family),
  return_lasso = TRUE,
  return_x_basis = FALSE,
  ...
)
}
\arguments{
\item{X}{An input \code{matrix} with dimensions number of observations -by-
number of covariates that will be used to derive the design matrix of basis
functions.}

\item{Y}{A \code{numeric} vector of observations of the outcome variable. For
\code{family="mgaussian"}, \code{Y} is a matrix of observations of the
outcome variables.}

\item{X_unpenalized}{An input \code{matrix} with the same number of rows as
\code{X}, for which no L1 penalization will be performed. Note that
\code{X_unpenalized} is directly appended to the design matrix; no basis
expansion is performed on \code{X_unpenalized}.}

\item{max_degree}{The highest order of interaction terms for which basis
functions ought to be generated.}

\item{smoothness_orders}{An \code{integer}, specifying the smoothness of the
basis functions. See details for \code{smoothness_orders} for more
information.}

\item{num_knots}{An \code{integer} vector of length 1 or \code{max_degree},
specifying the maximum number of knot points (i.e., bins) for any covariate
for generating basis functions. If \code{num_knots} is a unit-length
vector, then the same \code{num_knots} are used for each degree (this is
not recommended). The default settings for \code{num_knots} are
recommended, and these defaults decrease \code{num_knots} with increasing
\code{max_degree} and \code{smoothness_orders}, which prevents (expensive)
combinatorial explosions in the number of higher-degree and higher-order
basis functions generated. This allows the complexity of the optimization
problem to grow scalably. See details of \code{num_knots} more information.}

\item{reduce_basis}{Am optional \code{numeric} value bounded in the open
unit interval indicating the minimum proportion of 1's in a basis function
column needed for the basis function to be included in the procedure to fit
the lasso. Any basis functions with a lower proportion of 1's than the
cutoff will be removed. Defaults to 1 over the square root of the number of
observations. Only applicable for models fit with zero-order splines, i.e.
\code{smoothness_orders = 0}.}

\item{family}{A \code{character} or a \code{\link[stats]{family}} object
(supported by \code{\link[glmnet]{glmnet}}) specifying the error/link
family for a generalized linear model. \code{character} options are limited
to "gaussian" for fitting a standard penalized linear model, "binomial" for
penalized logistic regression, "poisson" for penalized Poisson regression,
"cox" for a penalized proportional hazards model, and "mgaussian" for
multivariate penalized linear model. Note that passing in
family objects leads to slower performance relative to passing in a
character family (if supported). For example, one should set
\code{family = "binomial"} instead of \code{family = binomial()} when
calling \code{fit_hal}.}

\item{lambda}{User-specified sequence of values of the regularization
parameter for the lasso L1 regression. If \code{NULL}, the default sequence
in \code{\link[glmnet]{cv.glmnet}} will be used. The cross-validated
optimal value of this regularization parameter will be selected with
\code{\link[glmnet]{cv.glmnet}}. If \code{fit_control}'s \code{cv_select}
argument is set to \code{FALSE}, then the lasso model will be fit via
\code{\link[glmnet]{glmnet}}, and regularized coefficient values for each
lambda in the input array will be returned.}

\item{id}{A vector of ID values that is used to generate cross-validation
folds for \code{\link[glmnet]{cv.glmnet}}. This argument is ignored when
\code{fit_control}'s \code{cv_select} argument is \code{FALSE}.}

\item{weights}{observation weights; defaults to 1 per observation.}

\item{offset}{a vector of offset values, used in fitting.}

\item{fit_control}{List of arguments, including the following, and any
others to be passed to \code{\link[glmnet]{cv.glmnet}} or
\code{\link[glmnet]{glmnet}}.
\itemize{
\item \code{cv_select}: A \code{logical} specifying if the sequence of
specified \code{lambda} values should be passed to
\code{\link[glmnet]{cv.glmnet}} in order for a single, optimal value of
\code{lambda} to be selected according to cross-validation. When
\code{cv_select = FALSE}, a \code{\link[glmnet]{glmnet}} model will be
used to fit the sequence of (or single) \code{lambda}.
\item \code{use_min}: Specify the choice of lambda to be selected by
\code{\link[glmnet]{cv.glmnet}}. When \code{TRUE}, \code{"lambda.min"} is
used; otherwise, \code{"lambda.1se"}. Only used when
\code{cv_select = TRUE}.
\item \code{lambda.min.ratio}: A \code{\link[glmnet]{glmnet}} argument
specifying the smallest value for \code{lambda}, as a fraction of
\code{lambda.max}, the (data derived) entry value (i.e. the smallest value
for which all coefficients are zero). We've seen that not setting
\code{lambda.min.ratio} can lead to no \code{lambda} values that fit the
data sufficiently well.
\item \code{prediction_bounds}: An optional vector of size two that provides
the lower and upper bounds predictions; not used when
\code{family = "cox"}. When \code{prediction_bounds = "default"}, the
predictions are bounded between \code{min(Y) - sd(Y)} and
\code{max(Y) + sd(Y)} for each outcome (when \code{family = "mgaussian"},
each outcome can have different bounds). Bounding ensures that there is
no extrapolation.
}}

\item{screener_control}{A list of parameters for the earth-based screening algorithm.
Options includee
-- \code{screen_variables}: A \code{logical} of whether to screen variables using MARS-based screening as implemented in \code{screen_MARS}.
If \code{TRUE}, then the selectively adaptive lasso routine \code{fit_hal_with_screening} is called internally.
Note that \code{fit_hal} may be much slower if this is set to \code{FALSE}
-- \code{screen_interactions}:  A \code{logical} of whether to screen interactions using MARS-based screening as implemented in \code{screen_MARS}.
Only used if \code{screen_variables} is \code{TRUE}.
This argument is passed to \code{fit_hal_with_screening} internally if \code{screen_variables} is \code{TRUE}.
Note that \code{fit_hal} may be slower if this is set to \code{FALSE}.
-- \code{max_degree}: Only used if \code{screen_variables} is \code{TRUE} and \code{screen_interactions} is \code{FALSE}.
The maximum degree of interaction to search for in the MARS-based selectively adaptive
lasso routine as implemented in \code{fit_hal_with_screening}.
-- \code{pruning_method}: Only used if \code{screen_variables} is \code{TRUE}.
The pruning method to select the MARS-based variable and interaction screener.
NOTE that HAL uses honest cross-validation and is thus robust to
the aggressiveness of the MARS-based fitting algorithm used for screening.
See the \code{pmethod} argument of \code{\link[earth]{earth}}.
The option \code{cv} uses 10-fold cross-validation (CV).
The option \code{backward} and \code{forward} prunes using backward and forward selection with the generalized cross-validation criteria (GCV).
GCV is a penalty that penalizes model complexity/size and is an approximation of leave-one-out CV.
The option \code{backward} avoids cross-validation and can thus be substantially faster than \code{cv}.
GCV-based pruning methods tend to select more variables and interactions than CV and
may be preferred in larger sample sizes.
-- \code{family}: A \code{\link[stats]{family}} object that is passed to \code{\link[earth]{earth}} during screening.
By default, \code{family} is the same as \code{family}.
However, \code{\link[earth]{earth}} may have convergence issues and/or error when \code{family} is not \code{"gaussian"}..}

\item{return_lasso}{A \code{logical} indicating whether or not to return
the \code{\link[glmnet]{glmnet}} fit object of the lasso model.}

\item{return_x_basis}{A \code{logical} indicating whether or not to return
the matrix of (possibly reduced) basis functions used in \code{fit_hal}.}
}
\value{
Object of class \code{hal9001}, containing a list of basis
functions, a copy map, coefficients estimated for basis functions, and
timing results (for assessing computational efficiency).
}
\description{
Estimation procedure for HAL with MARS-based screening
}
\details{
The procedure implements a greedy version of HAL that uses the
multivariate adaptive regression splines (MARS) implementation of \code{\link[earth]{earth}}
for variable selection and variable subgroup selection for HAL interactions.
By using earth to learn the structural form of the regression in a greedy manner,
the selective HAL is able to run much faster than standard HAL
and provides relatively quick solutions in large samples and moderately high dimensions.
}
\examples{
n <- 100
p <- 2
x <- xmat <- matrix(rnorm(n * p), n, p)
colnames(x) <- paste0("X", 1:p)
y_prob <- plogis(sin(x[, 1]) + sin(x[, 2]))
y <- rbinom(n = n, size = 1, prob = y_prob)
hal_fit <- fit_hal_with_screening(X = x, Y = y, family = "binomial", max_degree = 1, num_knots = 10)
print(hal_fit$formula)
preds <- predict(hal_fit, new_data = x)
}
