% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_hal9001.R
\name{fill_dots}
\alias{fill_dots}
\title{Formula Helpers}
\usage{
fill_dots(var_names, .)
}
\arguments{
\item{var_names}{A \code{character} vector of variable names representing a single type of interaction
var_names may include the wildcard variable "." in which case the argument \code{.} must be specified
so that all interactions matching the form of var_names are generated.}

\item{.}{Specification of variables for use in the formula.
This function takes a character vector \code{var_names} of the form c(name1, name2, ".", name3, ".")
with any number of name{int} variables and any number of wild card variables ".".
It returns a list of character vectors of the form c(name1, name2, wildcard1, name3, wildcard2)
where wildcard1 and wildcard2 are iterated over all possible character names given in the argument \code{.}.}
}
\description{
Formula Helpers
}
