% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{functional_mean}
\alias{functional_mean}
\title{Calculate the Mean Prediction of a HAL Fit}
\usage{
functional_mean(
  hal_fit,
  X,
  X_unpenalized = NULL,
  offset = NULL,
  other_data = NULL,
  ...
)
}
\arguments{
\item{hal_fit}{A HAL fit object.}

\item{X}{Data frame or matrix of predictors.}

\item{X_unpenalized}{An optional matrix of unpenalized predictors, if different from those
used in the original model fitting.}

\item{offset}{An optional vector of offset values to be used in the prediction.}

\item{other_data}{An optional data frame or matrix containing additional data used in prediction.}
}
\value{
The mean of the predicted values.
}
\description{
This utility function calculates the mean prediction from a HAL model, which can be used
as a functional in various inference procedures.
}
\examples{
# Assuming `hal_fit` is a HAL fit object
mean_pred <- functional_mean(hal_fit, X = my_data)

}
