% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{generate_all_rules}
\alias{generate_all_rules}
\title{Generates rules based on knot points of the fitted HAL basis functions with
non-zero coefficients.}
\usage{
generate_all_rules(basis_list, coefs, X_colnames)
}
\description{
Generates rules based on knot points of the fitted HAL basis functions with
non-zero coefficients.
}
\keyword{internal}
