% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_new_beta}
\alias{get_new_beta}
\title{Compute updated LASSO coefficients}
\usage{
get_new_beta(X, resids, j, beta_j, xscale_j)
}
\arguments{
\item{X}{...}

\item{resids}{...}

\item{j}{...}

\item{beta_j}{...}

\item{xscale_j}{...}
}
\description{
Compute updated LASSO coefficients
}
