% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9000.R
\name{hal9000}
\alias{hal9000}
\title{HAL 9000 Quotes}
\usage{
hal9000(seed = 79246)
}
\arguments{
\item{seed}{Passed directly to \code{set.seed}.}
}
\description{
Prints a quote from the HAL 9000 robot from 2001: A Space Odyssey
}
