% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal_quotes.R
\docType{data}
\name{hal_quotes}
\alias{hal_quotes}
\title{HAL 9000 Quotes from 2001: A Space Odyssey}
\format{A vector of quotes.}
\usage{
hal_quotes
}
\description{
Curated quotes from the HAL 9000 computer, from the 1968 film 2001: A Space
Odyssey.
}
\keyword{datasets}
