% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_first_copy}
\alias{index_first_copy}
\title{Find Copies of Columns}
\usage{
index_first_copy(X)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}
}
\description{
Index vector that, for each column in X, indicates the index of the first
copy of that column
}
