% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_delta_method}
\alias{inference_delta_method}
\title{Perform Inference Using the Delta Method with Bootstrapped HAL Fits}
\usage{
inference_delta_method(
  hal_fit,
  functional,
  alpha = 0.05,
  other_data = NULL,
  bootstrap_other_data = TRUE
)
}
\arguments{
\item{hal_fit}{A HAL fit object that has been bootstrapped using \code{bootstrap_hal}.}

\item{functional}{A function that takes the parameters \code{hal_fit}, \code{X}, \code{X_unpenalized}, \code{offset},
and \code{other_data}, and returns the calculated functional of interest. This function must output
a one-dimensional matrix representing the estimated values.}

\item{alpha}{The significance level used to construct the confidence interval (default is 0.05).}

\item{other_data}{An optional data frame or matrix containing additional data used in the functional.
If \code{other_data} is a vector, it is automatically converted to a matrix.}

\item{bootstrap_other_data}{A logical value indicating whether to bootstrap \code{other_data} alongside
the main data. If \code{TRUE}, \code{other_data} is resampled according to the bootstrap indices (default is TRUE).}
}
\value{
A matrix where each row corresponds to an estimate from the \code{functional} along with its
lower and upper confidence limits.
}
\description{
This function applies the delta method for inference using bootstrapped HAL fits.
It requires that the HAL model has been previously bootstrapped using \code{bootstrap_hal}.
It supports one-dimensional outcomes and allows for additional data to be included in the inference process.
}
\note{
Before using this function, ensure that the HAL fit has been bootstrapped by running
\code{hal_fit_bootstrapped <- bootstrap_hal(hal_fit)}.
}
\examples{
# Assuming `hal_fit_bootstrapped` is a previously bootstrapped HAL fit object
# and `my_functional` is defined to take the required arguments.
results <- inference_delta_method(hal_fit_bootstrapped, my_functional)

}
\seealso{
\code{\link{bootstrap_hal}} for details on bootstrapping HAL models.
}
