% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference_pointwise}
\alias{inference_pointwise}
\title{Provide Pointwise Confidence Intervals for Regression Fits by HAL}
\usage{
inference_pointwise(
  hal_fit,
  new_data,
  new_X_unpenalized = NULL,
  offset = NULL,
  alpha = 0.05
)
}
\arguments{
\item{hal_fit}{A previously bootstrapped HAL fit object.}

\item{new_data}{Data frame or matrix of new data for prediction.}

\item{new_X_unpenalized}{An optional matrix of unpenalized predictors, if different from those
used in the original model fitting.}

\item{offset}{An optional vector of offset values to be used in the prediction.}

\item{alpha}{The significance level used to construct the confidence interval (default is 0.05).}
}
\value{
A matrix where each row corresponds to a prediction from the new data along with its
lower and upper confidence intervals.
}
\description{
This function calculates pointwise confidence intervals for predictions from a HAL model
that has been bootstrapped. It is essential that the HAL model is bootstrapped prior to using
this function.
}
\note{
Ensure that the HAL fit has been bootstrapped by running:
\code{hal_fit_bootstrapped <- bootstrap_hal(hal_fit)}.
}
\examples{
# Assuming `hal_fit_bootstrapped` is a previously bootstrapped HAL fit object
results <- inference_pointwise(hal_fit_bootstrapped, new_data = my_new_data)

}
\seealso{
\code{\link{bootstrap_hal}} for details on how to bootstrap HAL models.
}
