% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{kmeans_knot_screen}
\alias{kmeans_knot_screen}
\title{Screen knot-points using k-means clustering}
\usage{
kmeans_knot_screen(
  X,
  bins,
  degree,
  smoothness_orders,
  include_zero_order,
  include_lower_order
)
}
\arguments{
\item{X}{A \code{numeric} vector or matrix of input.}

\item{bins}{A \code{numeric} scalar indicating the number of bins into which
\code{X} should be discretized.}
}
\description{
Screen knot-points using k-means clustering
}
\keyword{internal}
