% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassi.R
\name{lambda_seq}
\alias{lambda_seq}
\title{Generate sequence of lambdas}
\usage{
lambda_seq(lambda_max, lambda_min_ratio = 0.01, nlambda = 100)
}
\arguments{
\item{lambda_max}{the highest lambda value, ideally from get_lambda_max}

\item{lambda_min_ratio}{the ratio of the largest to smallest lambda values
lambda_min = lambda_max * lambda_min_ratio}

\item{nlambda}{total number of lambdas}
}
\description{
Generate sequence of lambdas
}
