% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassi.R
\name{lassi}
\alias{lassi}
\title{custom lasso implementation for indicators}
\usage{
lassi(x, y, nlambda = 100, lambda_min_ratio = 0.01)
}
\arguments{
\item{x}{the covariate matrix}

\item{y}{the outcome vector}

\item{nlambda}{number of lambdas to fit. See \code{\link{lambda_seq}}}

\item{lambda_min_ratio}{ratio of largest to smallest lambda to fit. See \code{\link{lambda_seq}}}
}
\description{
custom lasso implementation for indicators
}
