% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lassi_fit_cd}
\alias{lassi_fit_cd}
\title{Fit a LASSO Regression Model}
\usage{
lassi_fit_cd(X, y, lambda, nsteps)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}

\item{y}{Numeric containing observations of an outcome variable of interest.}

\item{lambda}{Numeric corresponding to the LASSO regularization parameter.}

\item{nsteps}{Maximum number of steps to take until stopping computation of
the regression coefficient.}
}
\description{
Fit a linear regression model with L1 penalization, the LASSO
}
