% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassi.R
\name{lassi_fit_module
Custom Lasso implementation for matrices of indicator functions}
\alias{lassi_fit_module
Custom Lasso implementation for matrices of indicator functions}
\alias{lassi}
\title{Rcpp module: lassi_fit_module}
\usage{
lassi(x, y, lambdas = NULL, nlambda = 100, lambda_min_ratio = 0.01,
  center = FALSE)
}
\arguments{
\item{x}{The covariate matrix}

\item{y}{The outcome vector}

\item{lambdas}{A sequence of values for the L1 regularization parameter
(lambda) to be used in fitting the LASSO. Defaults to \code{NULL}.}

\item{nlambda}{number of lambdas to fit. See \code{\link{lambda_seq}}}

\item{lambda_min_ratio}{ratio of largest to smallest lambda to fit. For
details, see \code{\link{lambda_seq}}}

\item{center}{...}
}
\description{
Rcpp module: lassi_fit_module
}
