% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{lassi_origami}
\alias{lassi_origami}
\title{Single LASSO estimation for cross-validation with Origami}
\usage{
lassi_origami(fold, data, lambdas)
}
\arguments{
\item{fold}{...}

\item{data}{...}

\item{lambdas}{...}
}
\description{
Fits the LASSO regression over a single fold of a cross-validated data set.
This is meant to be called using \code{cross_validate} from the
\code{origami} package, which is done through \code{cv_lasso}. Note that this
procedure is NOT meant to be invoked by itself. INTERNAL USE ONLY.
}
