% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lassi_predict}
\alias{lassi_predict}
\title{Prediction from a Lassi Model}
\usage{
lassi_predict(X, beta, intercept)
}
\arguments{
\item{X}{A sparse matrix of HAL basis functions.}

\item{beta}{A vector of coefficient values for the HAL basis functions.}

\item{intercept}{A numeric value giving the intercept of the HAL model.}
}
\description{
Prediction from a Lassi Model
}
