% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dupes.R
\name{make_copy_map}
\alias{make_copy_map}
\title{Build Copy Maps}
\usage{
make_copy_map(x_basis)
}
\arguments{
\item{x_basis}{A design matrix consisting of basis (indicator) functions for
covariates (X) and terms for interactions thereof.}
}
\description{
Build Copy Maps
}
