% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_design_matrix}
\alias{make_design_matrix}
\title{Build HAL Design Matrix}
\usage{
make_design_matrix(X, blist)
}
\arguments{
\item{X}{Matrix of covariates containing observed data in the columns.}

\item{blist}{List of basis functions with which to build HAL design matrix.}
}
\description{
Make a HAL design matrix based on original design matrix X and a list of
basis functions in blist
}
