% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_basis_filter.R
\name{make_reduced_basis_map}
\alias{make_reduced_basis_map}
\title{TITLE}
\usage{
make_reduced_basis_map(x_basis, keep_prop)
}
\arguments{
\item{x_basis}{...}

\item{keep_prop}{...}
}
\description{
description
}
