% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{num_knots_generator}
\alias{num_knots_generator}
\title{A default generator for the num_knots argument for each degree of interactions
and the smoothness orders.}
\usage{
num_knots_generator(d, smoothness_orders)
}
\description{
A default generator for the num_knots argument for each degree of interactions
and the smoothness orders.
}
