% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{pam_knot_screen}
\alias{pam_knot_screen}
\title{Screen knot-points using PAM clustering}
\usage{
pam_knot_screen(
  X,
  bins,
  degree,
  smoothness_orders,
  include_zero_order,
  include_lower_order
)
}
\arguments{
\item{X}{A \code{numeric} vector or matrix of input.}

\item{bins}{A \code{numeric} scalar indicating the number of knot-points for
each basis function.}

\item{degree}{A \code{numeric} scalar indicating the degree of the basis
functions.}

\item{smoothness_orders}{Argument for \code{enumerate_basis()}}

\item{include_zero_order}{Argument for \code{enumerate_basis()}}

\item{include_lower_order}{Argument for \code{enumerate_basis()}}
}
\description{
Using PAM clustering at each basis function level to screen
knot-points. An effective way to reduce the number of basis function and
improve the computational speed while maintaining good predictive
performance of HAL fit.
}
\keyword{internal}
