% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_hal9001.R
\name{+.formula_hal9001}
\alias{+.formula_hal9001}
\title{HAL Formula addition: Adding formula term object together into a single
formula object term.}
\usage{
\method{+}{formula_hal9001}(x, y)
}
\arguments{
\item{x}{A \code{formula_hal9001} object as outputted by \code{h}.}

\item{y}{A \code{formula_hal9001} object as outputted by \code{h}.}
}
\description{
HAL Formula addition: Adding formula term object together into a single
formula object term.
}
