% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassi.R
\name{predict.lassi}
\alias{predict.lassi}
\title{Predict Method for Lasso on Indicator Bases}
\usage{
\method{predict}{lassi}(fit, new_x_basis, lambdas = NULL)
}
\arguments{
\item{fit}{...}

\item{new_x_basis}{...}

\item{lambdas}{...}
}
\description{
Predict Method for Lasso on Indicator Bases
}
\keyword{internal}
