% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mthal9001}
\alias{predict.mthal9001}
\title{Prediction from MT-HAL fits}
\usage{
\method{predict}{mthal9001}(object, new_data)
}
\arguments{
\item{object}{An object of class \code{mthal9001}, containing the results of
fitting the Multi-task Highly Adaptive Lasso, as produced by
\code{\link{fit_mthal}}.}

\item{new_data}{A \code{numeric} input \code{matrix} that the
\code{mthal9001} object will use to derive the predicted values.}
}
\value{
A \code{matrix} of predictions from a \code{mthal9001} object. The
predictions will be truncated if in the call to \code{fit_mthal}, either
(a) \code{prediction_bounds} were supplied or (b) \code{type = "Regression"}
and \code{prediction_bounds} were set to default, and if the generated
predictions fall outside of those bounds.
}
\description{
Prediction from MT-HAL fits
}
\details{
Method for computing and extracting predictions from fits of the
Multi-task Highly Adaptive Lasso estimator, returned as a single S3 objects
of class \code{mthal9001}.
}
