% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.hal9001.R
\name{predict.SL.hal9001}
\alias{predict.SL.hal9001}
\title{predict.SL.hal9001}
\usage{
\method{predict}{SL.hal9001}(object, newdata)
}
\arguments{
\item{object}{A fitted object of class \code{hal9001}.}

\item{newdata}{A matrix of new observations on which to obtain predictions.}
}
\description{
Predict method for objects of class \code{SL.hal9001}
}
