% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_hal9001.R
\name{print.formula_hal9001}
\alias{print.formula_hal9001}
\title{Print formula_hal9001 object}
\usage{
\method{print}{formula_hal9001}(x, ...)
}
\arguments{
\item{x}{A formula_hal9001 object.}

\item{...}{Other arguments (ignored).}
}
\description{
Print formula_hal9001 object
}
