% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.hal9001}
\alias{print.summary.hal9001}
\title{Print Method for Summary Class of HAL fits}
\usage{
\method{print}{summary.hal9001}(x, length = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{summary.hal9001}.}

\item{length}{The number of ranked coefficients to be summarized.}

\item{...}{Other arguments (ignored).}
}
\description{
Print Method for Summary Class of HAL fits
}
