% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hal_with_screening.R, R/screening.R
\name{safe_dim}
\alias{safe_dim}
\title{Borrowed from sl3. Internal use.}
\usage{
safe_dim(x)

safe_dim(x)
}
\description{
Borrowed from sl3. Internal use.

Borrowed from sl3. Internal use.
}
