% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Soft thresholding for LASSO fits}
\usage{
soft_threshold(beta, lambda)
}
\arguments{
\item{beta}{Numeric of the regression coefficients of a linear model.}

\item{lambda}{Numeric of the regularization constant for the L1 penalty.}
}
\description{
The soft thresholding algorithm given by Hastie et al. (2009)
}
