% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{X_t_resid}
\alias{X_t_resid}
\title{Coordinate Descent with Regression Residuals}
\usage{
X_t_resid(X, resids, j, xscale_j, xcenter_j, resid_sum)
}
\arguments{
\item{X}{A sparse matrix corresponding to the full set of basis functions.}

\item{resids}{Residuals from a previous round of regression fits.}

\item{j}{A numeric value indexing the columns of X (i.e., basis functions).}

\item{xscale_j}{A numeric corresponding to the scaled values of the jth
  column of the sparse matrix X.

compute $X'r$ for a given column of X}
}
\description{
Coordinate Descent with Regression Residuals
}
