% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_copy_map}
\alias{apply_copy_map}
\title{Apply copy map}
\usage{
apply_copy_map(X, copy_map)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}

\item{copy_map}{the copy map}
}
\description{
OR duplicate training set columns together
}
