% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{basis_list_cols}
\alias{basis_list_cols}
\title{List Basis Functions}
\usage{
basis_list_cols(cols, x, order_map, include_zero_order)
}
\arguments{
\item{cols}{Index or indices (as \code{numeric}) of covariates (columns) of
interest in the data matrix \code{x} for which basis functions ought to be
generated. Note that basis functions for interactions of these columns are
computed automatically.}

\item{x}{A \code{matrix} containing observations in the rows and covariates
in the columns. Basis functions are computed for these covariates.}
}
\value{
A \code{list} containing the basis functions generated from a set of
 input columns.
}
\description{
Build a list of basis functions from a set of columns
}
