% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{enumerate_basis}
\alias{enumerate_basis}
\title{Enumerate Basis Functions}
\usage{
enumerate_basis(
  x,
  max_degree = NULL,
  order_map = rep(0, ncol(x)),
  include_zero_order = F,
  include_lower_order = F,
  num_bins = NULL
)
}
\arguments{
\item{x}{An input \code{matrix} containing observations and covariates
following standard conventions in problems of statistical learning.}

\item{max_degree}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{order_map}{An integer vector of length ncol(x) specifying the desired smoothness of the function in each covariate.
 k = 0 is no smoothness (indicator basis), k = 1 is first order smoothness, and so on.
For an additive model, the component function for each covariate will have the degree of smoothness as specified by order_map.
For non-additive components (tensor products of single variate basis functions),
the univariate basis functions in each tensor product have smoothness degree as specified by order_map.}

\item{include_zero_order}{A boolean variable. If true the zero order basis functions are included for each covariate, in addition to the smooth basis functions specified by order_map.
This allows the algorithm to data-adaptively choose the smoothness.}

\item{include_lower_order}{A boolean variable. Similar to include_zero_order, except inclues all basis functions of lower smoothness degrees than specified via order_map.}

\item{num_bins}{A vector of length max_degree which determines how granular the knots points to generate basis functions for should be for each degree of basis function.
num_bins[1] determines the number of knot points to be used for each univariate basis function.
More generally, num_bins[k] determines the number of knot points to be used for the kth degree basis functions.
Specifically, for a kth degree basis function, which is the tensor product of k univariate basis functions, ...
... this determines the number of knot points to be used for each univariate basis function in the tensor product}
}
\value{
A \code{list} of basis functions generated for all covariates and
 interaction thereof up to a pre-specified degree.
}
\description{
Generate basis functions for all covariates and interaction terms thereof up
to a specified order/degree
}
\examples{
\donttest{
gendata <- function(n) {
  W1 <- runif(n, -3, 3)
  W2 <- rnorm(n)
  W3 <- runif(n)
  W4 <- rnorm(n)
  g0 <- plogis(0.5 * (-0.8 * W1 + 0.39 * W2 + 0.08 * W3 - 0.12 * W4))
  A <- rbinom(n, 1, g0)
  Q0 <- plogis(0.15 * (2 * A + 2 * A * W1 + 6 * A * W3 * W4 - 3))
  Y <- rbinom(n, 1, Q0)
  data.frame(A, W1, W2, W3, W4, Y)
}
set.seed(1234)
data <- gendata(100)
covars <- setdiff(names(data), "Y")
X <- as.matrix(data[, covars, drop = FALSE])
basis_list <- enumerate_basis(X)
}

}
