% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{fit_hal}
\alias{fit_hal}
\title{HAL: The Highly Adaptive LASSO estimator}
\usage{
fit_hal(X, Y, degrees = NULL, yolo = TRUE, useMin = TRUE, ...)
}
\arguments{
\item{X}{An input \code{matrix} containing observations and covariates
following standard conventions in problems of statistical learning.}

\item{Y}{A \code{numeric} vector of obervations of the outcome variable of
interest, following standard conventions in problems of statistical learning.}

\item{degrees}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{yolo}{A \code{logical} indicating whether to print one of a curated
selection of quotes from HAL 9000, from 2001: A Space Odyssey (1968).}

\item{useMin}{Determines which lambda is selected from \code{cv.glmnet}. True means \code{"lambda.min"} is used, otherwise \code{"lambda.1se"}}

\item{...}{Other arguments passed to \code{cv.glmnet}. Please consult the
documentation for \code{glmnet} for a full list of options.}
}
\value{
Object of class \code{hal9001}, containing a list of basis functions,
a copy map, coefficients estimated for basis functions, and timing results
(for assessing computational efficiency).
}
\description{
Performs the estimation procedure for HAL, the Highly Adaptive LASSO
}
\details{
The procedure uses custom C++ functions to generate the design
matrix (consisting of basis functions corresponding to covariates and
interactions of covariates) and remove duplicate columns of indicators. The
actual LASSO regression that follows is computed via \code{cv.glmnet},
though plans are in place to re-implement this in Rcpp/C++ as well.
}
