% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{fit_hal}
\alias{fit_hal}
\title{HAL: The Highly Adaptive LASSO estimator}
\usage{
fit_hal(X, Y, degrees = NULL)
}
\arguments{
\item{X}{...}

\item{Y}{...}

\item{degrees}{...}
}
\value{
Object of class \code{hal9001}, containing ...
}
\description{
Performs the estimation procedure of the Highly Adaptive LASSO (HAL)
}
\details{
The procedure uses custom C++ functions to generate the design
matrix (consisting of basis functions corresponding to covariates and
interactions of covariates) and remove duplicate columns of indicators. The
actual LASSO regression that follows is computed via \code{cv.glmnet},
though plans are in place to re-implement this in Rcpp/C++ as well.
}
\examples{

}
