% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal.R
\name{fit_hal}
\alias{fit_hal}
\title{HAL: The Highly Adaptive LASSO estimator}
\usage{
fit_hal(X, Y, degrees = NULL, fit_type = c("origami", "glmnet"),
  n_folds = 10, use_min = TRUE, ..., yolo = TRUE)
}
\arguments{
\item{X}{An input \code{matrix} containing observations and covariates
following standard conventions in problems of statistical learning.}

\item{Y}{A \code{numeric} vector of obervations of the outcome variable of
interest, following standard conventions in problems of statistical learning.}

\item{degrees}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{fit_type}{The specific routine to be called when fitting the LASSO
regression in a cross-validated manner. Choosing the \code{glmnet} option
will result in a call to \code{cv.glmnet} while \code{origami} will produce
a (faster) call to a custom routine based on the \code{origami} package.}

\item{n_folds}{Integer for the number of folds to be used when splitting the
data for cross-validation. This defaults to 10 as this is the convention for
v-fold cross-validation.}

\item{use_min}{Determines which lambda is selected from \code{cv.glmnet}.
\code{TRUE} corresponds to \code{"lambda.min"} and \code{FALSE} corresponds
to \code{"lambda.1se"}.}

\item{...}{Other arguments passed to \code{cv.glmnet}. Please consult the
documentation for \code{glmnet} for a full list of options.}

\item{yolo}{A \code{logical} indicating whether to print one of a curated
selection of quotes from the HAL9000 computer, from the critically acclaimed
epic science-fiction film "2001: A Space Odyssey" (1968).}
}
\value{
Object of class \code{hal9001}, containing a list of basis functions,
 a copy map, coefficients estimated for basis functions, and timing results
 (for assessing computational efficiency).
}
\description{
Estimation procedure for HAL, the Highly Adaptive LASSO
}
\details{
The procedure uses a custom C++ implementation to generate a design
 matrix (consisting of basis functions corresponding to covariates and
 interactions of covariates) and remove duplicate columns of indicators. The
 LASSO regression is fit to this (usually) very wide matrix using either a
 custom implementation (based on the \code{origami} package) or by a call to
 \code{cv.glmnet} from the \code{glmnet} package.
}
