% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal_cate.R
\name{fit_hal_cate}
\alias{fit_hal_cate}
\title{CATE-HAL: Highly Adaptive Lasso for CATE Estimation}
\usage{
fit_hal_cate(
  X,
  Y,
  A,
  A.hat,
  Y.hat,
  A_fit_params = NULL,
  Y_fit_params = NULL,
  formula = NULL,
  X_unpenalized = NULL,
  max_degree = ifelse(ncol(X) >= 20, 2, 3),
  smoothness_orders = 1,
  num_knots = num_knots_generator(max_degree = max_degree, smoothness_orders =
    smoothness_orders, base_num_knots_0 = 200, base_num_knots_1 = 50),
  weights = NULL,
  ...
)
}
\arguments{
\item{X}{A \code{matrix} of covariates used to derive the design matrix of basis functions, with dimensions corresponding to the number of observations by the number of covariates.}

\item{Y}{A \code{numeric} vector containing observations of the outcome variable.}

\item{A}{A \code{numeric} vector containing observations of the binary treatment indicator.
Note that this vector must consist solely of 0's and 1's.}

\item{A.hat}{A \code{numeric} vector of estimates for the propensity score \code{E[A | X]},
which is the conditional mean of the treatment \code{A} given the covariates \code{X}.}

\item{Y.hat}{A \code{numeric} vector of estimates for \code{E[Y | X]},
which is the conditional mean of the outcome \code{Y} given the covariates \code{X}.}

\item{A_fit_params}{Optional parameters for fitting the model for the propensity score.}

\item{Y_fit_params}{Optional parameters for fitting the model for the outcome.}

\item{formula}{An optional formula specifying the model structure for the CATE estimation.}

\item{X_unpenalized}{An optional matrix of covariates that will not be penalized in the model.}

\item{max_degree}{The maximum degree of interactions considered in the model.
Default is set depending on the number of covariates: 2 if more than or equal to 20 covariates, otherwise 3.}

\item{smoothness_orders}{The order of smoothness applied to the basis functions, default is 1.}

\item{num_knots}{Function used to generate the number of knots for basis functions depending on the model complexity.}

\item{weights}{An optional vector of weights to be used in the regression.}

\item{...}{Additional arguments to be passed internally to \code{fit_hal}.
Note that the \code{family} argument should not be specified as it is internally set to "gaussian".
Specifying this argument through \code{...} will cause an argument clash error.}
}
\description{
This function estimates the Conditional Average Treatment Effect (CATE) using a Highly Adaptive Lasso (HAL)-based R-learner.
}
\details{
This procedure implements the R-learner approach (Nie & Wager, 2021) for estimating the Conditional Average Treatment Effect (CATE) with a binary treatment variable.
It utilizes the Highly Adaptive Lasso (HAL) estimator as implemented by the \code{fit_hal} function. This method functions similarly to \code{fit_hal}, but it requires additional arguments:
an initial estimator of the propensity score (the conditional mean of the treatment given the covariates) and the conditional mean of the outcome given the covariates.
}
