% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_hal9001.R
\name{formula_helpers}
\alias{formula_helpers}
\alias{fill_dots_helper}
\alias{fill_dots}
\title{Formula Helpers}
\usage{
fill_dots_helper(var_names, .)

fill_dots(var_names, .)
}
\arguments{
\item{var_names}{A \code{character} vector of variable names.}

\item{.}{Specification of variables for use in the formula.}
}
\description{
Formula Helpers
}
