% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal_quotes.R
\docType{data}
\name{hal_quotes}
\alias{hal_quotes}
\title{HAL9000 Quotes from "2001: A Space Odyssey"}
\format{
A vector of quotes.
}
\usage{
hal_quotes
}
\description{
Curated selection of quotes from the HAL9000 computer, from the critically
acclaimed epic science-fiction film "2001: A Space Odyssey" (1968).
}
\keyword{datasets}
