% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassi.R
\name{lassi}
\alias{lassi}
\title{Custom Lasso implementation for matrices of indicator functions}
\usage{
lassi(
  x,
  y,
  lambdas = NULL,
  nlambda = 100,
  lambda_min_ratio = 0.01,
  center = FALSE
)
}
\arguments{
\item{x}{The covariate matrix}

\item{y}{The outcome vector}

\item{lambdas}{A sequence of values for the L1 regularization parameter
(lambda) to be used in fitting the LASSO. Defaults to \code{NULL}.}

\item{nlambda}{number of lambdas to fit.}

\item{lambda_min_ratio}{ratio of largest to smallest lambda to fit.}

\item{center}{...}
}
\description{
Custom Lasso implementation for matrices of indicator functions
}
\keyword{internal}
