% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lassi_fit_cd}
\alias{lassi_fit_cd}
\title{Fit a LASSO Regression Model}
\usage{
lassi_fit_cd(X, resids, beta, lambda, nsteps, xscale, active_set)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}

\item{resids}{Numeric of the residuals from a given fit of the LASSO model.}

\item{beta}{Numeric vector of initial beta estiamtes}

\item{lambda}{Numeric corresponding to the LASSO regularization parameter.}

\item{nsteps}{Maximum number of steps to take until stopping computation of
the regression coefficient.}

\item{xscale}{scale factor for covariates. See get_xscale}

\item{active_set, }{update only nonzero coefficients (TRUE), or all
coefficients (FALSE)}
}
\description{
Fit a linear regression model with L1 penalization, the LASSO
}
