% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_lasso.R
\name{lassi_origami}
\alias{lassi_origami}
\title{Single LASSO estimation for cross-validation with Origami}
\usage{
lassi_origami(fold, data, lambdas)
}
\arguments{
\item{fold}{A \code{fold} object produced by a call to \code{make_folds} from
the \code{origami} package.}

\item{data}{A \code{dgCMatrix} object containing the outcome values (Y) in
its first column and vectors corresponding to the basis functions of HAL in
all other columns. Consult the description of the HAL algorithm for details.}

\item{lambdas}{A \code{numeric} vector corresponding to a sequence of lambda
values obtained by fitting the LASSO on the full data.}
}
\description{
Fits the LASSO regression over a single fold of a cross-validated data set.
This is meant to be called using \code{cross_validate} from the
\code{origami} package, which is done through \code{cv_lasso}. Note that this
procedure is NOT meant to be invoked by itself. INTERNAL USE ONLY.
}
