% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lassi_predict}
\alias{lassi_predict}
\title{LASSO Prediction}
\usage{
lassi_predict(X, beta)
}
\arguments{
\item{X}{Sparse matrix containing columns of indicator functions.}

\item{beta}{Numeric for the regression coefficient of a linear model fit.}
}
\description{
Compute predictions based on a LASSO regression
}
