% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{or_duplicate_columns}
\alias{or_duplicate_columns}
\title{Reorder Columns of X}
\usage{
or_duplicate_columns(X, cols)
}
\description{
ORs the columns of X listed in cols and places the result in column col[1]
}
