% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dupes.R
\name{os_find_dupes}
\alias{os_find_dupes}
\title{Finding Duplicate Columns (R)}
\usage{
os_find_dupes(X.init)
}
\arguments{
\item{X.init}{...}
}
\description{
Original HAL implementation of duplicate detection, for comparision purposes
}
\author{
Oleg Sofrygin
}
