% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dupes.R
\name{os_find_dupes}
\alias{os_find_dupes}
\title{Remove Duplicate Columns (R)}
\usage{
os_find_dupes(X.init)
}
\arguments{
\item{X.init}{...}
}
\description{
Duplicate column detection and removal, purely in R
}
\author{
Oleg Sofrygin
}
