% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.hal9001}
\alias{predict.hal9001}
\title{Prediction from HAL fits}
\usage{
\method{predict}{hal9001}(object, ..., newdata)
}
\arguments{
\item{object}{...}

\item{...}{Additional arguments passed to \code{predict} as necessary.}

\item{newdata}{...}
}
\description{
Prediction from HAL fits
}
\details{
<ethod for computing and extracting predictions from fits of the
Highly Adaptive LASSO estimator, formally S3 objects of class \code{hal9001}
}
\examples{

}
