% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_hal9001.R
\name{predict.SL.hal9001}
\alias{predict.SL.hal9001}
\title{predict.SL.hal9001}
\usage{
\method{predict}{SL.hal9001}(object, newX, ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal9001}.}

\item{newX}{A matrix of new observations on which to obtain predictions.}

\item{...}{Prevents unexpected process death. DON'T USE. (This is a hack.)}
}
\description{
Predict method for objects of class \code{SL.hal9001}
}
