% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_hal9001.R
\name{predict.SL.hal9001}
\alias{predict.SL.hal9001}
\title{predict.SL.hal9001}
\usage{
\method{predict}{SL.hal9001}(object, newdata, ...)
}
\arguments{
\item{object}{A fitted object of class \code{hal9001}.}

\item{newdata}{A matrix of new observations on which to obtain predictions.}

\item{...}{Placeholder (ignored).}
}
\value{
A \code{numeric} vector of predictions from a \code{SL.hal9001}
 object based on the provide \code{newdata}.
}
\description{
Predict method for objects of class \code{SL.hal9001}
}
