% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basis.R
\name{quantizer}
\alias{quantizer}
\title{Discretize Variables into Number of Bins by Unique Values}
\usage{
quantizer(X, bins)
}
\arguments{
\item{X}{A \code{numeric} vector to be discretized.}

\item{bins}{A \code{numeric} scalar indicating the number of bins into which
\code{X} should be discretized..}
}
\description{
Discretize Variables into Number of Bins by Unique Values
}
\keyword{internal}
