% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_with_mars.R
\name{screen_MARS}
\alias{screen_MARS}
\title{Screening variables and interactions with MARS (earth)}
\usage{
screen_MARS(
  x,
  y,
  pmethod = "cv",
  degree = 2,
  nfold = 10,
  fast.k = NULL,
  nk = NULL,
  glm = list(family = gaussian()),
  weights = NULL
)
}
\arguments{
\item{pmethod}{Model pruning method for earth. Default is to use cross-validation. See documentation of \code{\link[earth]{earth}}.}

\item{degree}{The max degree interaction of MARS model used for variable selection.
See documentation of \code{\link[earth]{earth}}.}

\item{nfold}{Number of folds for cross-validation. See documentation of \code{\link[earth]{earth}}.}

\item{glm}{A list of parameters to pass to \code{\link[earth]{glm}}.
See documentation of \code{\link[earth]{earth}}.}

\item{weights}{observation weights; defaults to 1 per observation.}
}
\value{
list contains variable selected and their column indices,
and a \code{hal_formula} object specifying the learned model.
}
\description{
Screening variables and interactions with MARS (earth)
}
\details{
A greedy procedure for learning variables and interaction subgroups for HAL using
multivariate adaptive regression splines (MARS) implementation of \code{\link[earth]{earth}}
}
\examples{
n <- 900
d <- 10
X <- replicate(d, runif(n))
colnames(X) <- paste0("X", 1:d)
mu <- 10 * sin(3 * X[, 5]) * sin(3 * X[, 3]) * sin(3 * X[, 1])
Y <- rnorm(n, mu, 0.5)
screen_MARS(X, Y, degree = 1)

}
