% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_hal9001.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{Wrapper for Classic SuperLearner}
\usage{
SL.hal9001(
  Y,
  X,
  newX = NULL,
  family = stats::gaussian(),
  obsWeights = rep(1, length(Y)),
  id = NULL,
  ...
)
}
\arguments{
\item{Y}{A \code{numeric} vector of observations of the outcome variable.}

\item{X}{A \code{matrix} of predictors/covariates.}

\item{newX}{A matrix of new observations on which to obtain predictions. The
default of \code{NULL} computes predictions on training inputs \code{X}.}

\item{family}{A \code{\link[stats]{family}} object (one that is supported
by \code{\link[glmnet]{glmnet}}) specifying the error/link family for a
generalized linear model.}

\item{obsWeights}{A \code{numeric} vector of observational-level weights.}

\item{...}{Additional arguments passed to \code{fit_hal}. See its
documentation for more details.}
}
\value{
An object of class \code{SL.hal9001} with a fitted \code{hal9001}
object and corresponding predictions based on the input data.
}
\description{
Wrapper for \pkg{SuperLearner} for objects of class \code{hal9001}
}
