% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_hal9001.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{Wrapper for Classic SuperLearner}
\usage{
SL.hal9001(Y, X, newX = NULL, degrees = NULL, family = stats::gaussian(),
  obsWeights = rep(1, length(Y)), ...)
}
\arguments{
\item{Y}{A \code{numeric} of outcomes.}

\item{X}{A \code{matrix} of predictors/covariates.}

\item{newX}{A matrix of new observations on which to obtain predictions. The
default of \code{NULL} computes predictions on training inputs \code{X}.}

\item{degrees}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{family}{Not used by the function directly, but ensures compatibility
with \code{SuperLearner}.}

\item{obsWeights}{Not used by the function directly, but meant to ensure
compatibility with \code{SuperLearner}.}

\item{...}{Prevents unexpected process death. DON'T USE. (This is a hack.)}
}
\description{
Wrapper for package \code{SuperLearner} for objects of class \code{hal9001}
}
