% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_hal9001.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{Classic SuperLearner Wrapper for HAL9001}
\usage{
SL.hal9001(X, Y, degrees = NULL, newdata = NULL, ...)
}
\arguments{
\item{X}{A \code{matrix} of predictors/covariates.}

\item{Y}{A \code{numeric} of outcomes.}

\item{degrees}{The highest order of interaction terms for which the basis
functions ought to be generated. The default (\code{NULL}) corresponds to
generating basis functions for the full dimensionality of the input matrix.}

\item{newdata}{A matrix of new observations on which to obtain predictions.}

\item{...}{Any other arguments to pass through to \code{hal9001::fit_hal},
which are then passed directly to \code{glmnet::cv.glmnet} (for now).}
}
\description{
Wrapper for package \code{SuperLearner} for objects of class \code{hal9001}
}
