% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squash_hal.R
\name{squash_hal_fit}
\alias{squash_hal_fit}
\title{Squash HAL objects}
\usage{
squash_hal_fit(object)
}
\arguments{
\item{object}{An object of class \code{hal9001}, containing the results of
fitting the Highly Adaptive LASSO, as produced by a call to \code{fit_hal}.}
}
\value{
Object of class \code{hal9001}, similar to the input object but
 reduced such that coefficients belonging to bases with coefficients equal to
 zero removed.
}
\description{
Reduce footprint by dropping basis functions with coefficients of zero
}
